Update seeds texture

+++ src/main/java/net/mcreator/starcraftvalley/procedures/BreakCheckProcedure.java
@@ -36,9 +36,9 @@ public class BreakCheckProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if (!((world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == TilledEarthBlock.block
-				|| (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getBlock() == WetEarthBlock.block)) {
-			world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
+		if (!((world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == TilledEarthBlock.block
+				|| (world.getBlockState(new BlockPos(x, y - 1, z))).getBlock() == WetEarthBlock.block)) {
+			world.destroyBlock(new BlockPos(x, y, z), false);
 		}
 	}
 }
